Added CSS to style the page.
Display Django messages near the top of the page.

Adjusted form action to include a query string with the file path.

Include the CSRF token.

Added a hidden input with the file path; needed to update the file
model.

Use form for any fields that can be edited instead of file.

Replaced file.size with file.human_readable_size to use largest
appropriate units.  Need to use a filter for this.

Use a more terse formatting for file.created_at.

Show the file's actual MD5 and SHA256 checksums.

Moved the footer to the bottom of the page.

Actually process the form when the update button is pressed.