Import more classes from .models.
Load the user's FileSummary and the FileQuota objects in
FileIndexView.get().

Specify the name of the file that was uploaded in FileIndexView.post()
when displaying an informational message.

If the form is not valid during FileIndexView.post() display detailed
error messages to the user.

Update the user's FileSummary object when processing a file deletion in
FileDeleteView.get().

Changed the file deletion message from an info to a warning.