Removed some print statements.
Get width and height settings from the query string and pass them along
when creating the SeleniumForm instance in SeleniumView.get().

Explicitly set the data keyword argument when constructing a
SeleniumForm in SeleniumView.post()

Build a query string containing width and height settings in
SeleniumView.post() to include with the redirect.